Fix code that looks like memory leak (#2367)
Creating a lua_ai_context puts a pointer to the ai::engine_lua to the Lua
stack, making Lua one of the owners of the engine. And on top of that,
the engine created the context implicitly.

This commit requires the caller to create the lua_ai_context instead. The
caller is always the initial owner of the engine, and therefore can give
ownership of the engine to Lua. In addition, functions which take ownership
of the engine now receive it as an rvalue reference to unique_ptr, to
ensure that we don't have owning raw pointers, which appear to confuse
Codacy.

Unfortunately, the engine isn't always singly owned. The AI itself manages
engines with shared pointers. Therefore I had to resort to the hack of
constructing a unique_ptr to an object that's already managed by a
shared_ptr.