Implemented config pooling
Unfortunately, this fails to improve performance. The main reason seems to
be that we keep child configs in a *map of vectors*, and therefore call the
destructors for those vectors whenever we clear a config, no matter whether
we call the destructor of the config itself. Therefore, most memory
allocations/deallocations aren't averted.

I changed attributes to be stored in a std::unordered_map for the same
reason. Std::map needs to free every element individually on clear, whereas
std::unordered_map may not need to, depending on how it's implemented.

Anyway, my performance measurements show that loading times are about on
par with master. Increasing complexity in exchange of nothing isn't exactly
a good deal.